Added a separate Service for transactional CallbackTaskExecutor methods.
